requirements.txt
PyQt5~=5.15.10
pandas~=2.2.2
matplotlib~=3.9.0
python-dotenv~=1.0.1
pygame~=2.5.2